renameSessions
rename instances of sessions to server, preparing for save actual
individual sessions feature
