+++ __tests__/sessions/loadSessions.test.ts
@@ -1,18 +1,18 @@
-import { loadLatestSession, handleSessionsIfServerIsRunning } from '@sessions/commands/loadSession';
-import * as sessionUtils from '@sessions/utils/sessionUtils';
+import { loadLatestSession, handleSessionsIfServerIsRunning } from '@tmux/commands/loadSession';
+import * as sessionUtils from '@tmux/utils/sessionUtils';
 import * as generalUI from '@UI/generalUI';
 import * as fs from 'fs/promises';
-import * as tmux from '@sessions/core/tmux';
+import * as tmux from '@tmux/core/tmux';
 import * as utils from '@utils/common';
-import { saveSessionsToFile } from '@sessions/commands/saveSessions';
+import { saveSessionsToFile } from '@tmux/commands/saveSessions';
 
-jest.mock('@sessions/utils/sessionUtils');
+jest.mock('@tmux/utils/sessionUtils');
 jest.mock('@UI/generalUI');
 jest.mock('fs/promises');
-jest.mock('@sessions/core/tmux');
+jest.mock('@tmux/core/tmux');
 jest.mock('@utils/common');
 jest.mock('@utils/neovimHelper');
-jest.mock('@sessions/commands/saveSessions');
+jest.mock('@tmux/commands/saveSessions');
 jest.mock('@utils/bashHelper', () => ({
   sendKeysToTmuxTargetSession: jest.fn(() => Promise.resolve())
 }));
@@ -115,7 +115,7 @@ describe('handleSessionsIfServerIsRunning', () => {
     it('should save sessions, kill server and sleep when confirmed', async () => {
         const currentSessions = { session1: { windows: [] } };
         mockGetCurrentSessions.mockResolvedValue(currentSessions);
-        jest.spyOn(require('@sessions/commands/printSessions'), 'printSessions').mockImplementation(mockPrintSessions);
+        jest.spyOn(require('@tmux/commands/printSessions'), 'printSessions').mockImplementation(mockPrintSessions);
         mockPrompt.mockResolvedValue(true);
         mockUtils.sleep.mockResolvedValue(undefined);
 
@@ -130,7 +130,7 @@ describe('handleSessionsIfServerIsRunning', () => {
     it('should only kill server when not confirmed to save sessions', async () => {
         const currentSessions = { session1: { windows: [] } };
         mockGetCurrentSessions.mockResolvedValue(currentSessions);
-        jest.spyOn(require('@sessions/commands/printSessions'), 'printSessions').mockImplementation(mockPrintSessions);
+        jest.spyOn(require('@tmux/commands/printSessions'), 'printSessions').mockImplementation(mockPrintSessions);
         mockPrompt.mockResolvedValue(false);
         mockUtils.sleep.mockResolvedValue(undefined);
