+++ auto.sh
@@ -16,7 +16,7 @@ _kra_completions() {
     if [[ "$prev" == "git" ]]; then
         COMPREPLY=( $(compgen -W "open-pr view-changed restore cache-untracked retrieve-untracked hard-reset log stash stash-drop-multiple conflict-handle" -- "$cur") )
     elif [[ "$prev" == "tmux" ]]; then
-        COMPREPLY=( $(compgen -W "save-server load-server list-sessions delete-session kill" -- "$cur") )
+        COMPREPLY=( $(compgen -W "save-server load-server list-sessions delete-server kill" -- "$cur") )
     elif [[ "$prev" == "sys" ]]; then
         COMPREPLY=( $(compgen -W "scripts grep-file-remove grep-dir-remove" -- "$cur") )
     elif [[ "$prev" == "ai" ]]; then
