--- src/sessions/index.ts
@@ -1,11 +1,0 @@
-// Core exports
-export * from '@sessions/core/tmux';
-
-// Command exports
-export * from '@sessions/commands/saveSessions';
-export * from '@sessions/commands/loadSession';
-export * from '@sessions/commands/manageSessions';
-export * from '@sessions/commands/printSessions';
-
-// Utility exports
-export * from '@sessions/utils/sessionUtils';
