+++ src/tmux/core/tmux.ts
