+++ src/tmux/index.ts
@@ -1,0 +1,11 @@
+// Core exports
+export * from '@tmux/core/tmux';
+
+// Command exports
+export * from '@tmux/commands/saveSessions';
+export * from '@tmux/commands/loadSession';
+export * from '@tmux/commands/manageSessions';
+export * from '@tmux/commands/printSessions';
+
+// Utility exports
+export * from '@tmux/utils/sessionUtils';
