+++ src/tmux/utils/formatters.ts
